Update typedd.rst
Re `cast` vs. `stringToNatOrZ` in the chapter 5 tip:

`cast` from `String` to `Nat` does work in Idris 2, version 0.6.0-a75161cb2, so the first sentence is not currently correct, and there is no change needed to make the code work.

In the Discord documentation channel, z_snail suggested changing the text to merely recommend changing `cast` to `stringToNatOrZ`.   This PR proposes that change.  I tried to change the text as little as possible, but I made some additional, strictly speaking unnecessary revisions, where I thought they would make the wording clearer given the essential changes.  I added a missing period as well.

(If the `cast` from `String` to `Nat` is destined to be removed, perhaps this PR should be ignored.)